move components to lib/
