+++ package.json
@@ -16,6 +16,7 @@
         "@mui/material": "^5.15.10",
         "@mui/material-nextjs": "^5.15.9",
         "@tanstack/react-query": "^5.20.2",
+        "@types/swagger-ui-react": "^4.18.3",
         "better-sqlite3": "^9.2.2",
         "kysely": "^0.27.1",
         "next": "14.1.1-canary.55",
@@ -23,7 +24,8 @@
         "radash": "^11.0.0",
         "react": "^18",
         "react-dom": "^18",
-        "react-hook-form": "^7.48.2"
+        "react-hook-form": "^7.48.2",
+        "swagger-ui-react": "^5.11.8"
     },
     "devDependencies": {
         "@types/better-sqlite3": "^7.6.8",
