+++ src/app/docs/page.tsx
@@ -1,0 +1,6 @@
+import "swagger-ui-react/swagger-ui.css"
+
+export default function Docs() {
+    // return <SwaggerUI url="https://petstore.swagger.io/v2/swagger.json" />
+    return "docs"
+}
