+++ src/lib/components/home/details/copy-link-button.tsx
