+++ src/lib/components/home/details/details-container.tsx
