+++ src/lib/components/home/details/details-content.tsx
