+++ src/lib/components/home/details/details-skeleton.tsx
