+++ src/lib/components/home/details/details.module.scss
