+++ src/lib/components/home/details/details.tsx
@@ -2,7 +2,6 @@ import { JobData } from "@/lib/job-data"
 import LaunchIcon from "@mui/icons-material/Launch"
 import { Button, Card, Paper } from "@mui/material"
 import { useEffect, useRef } from "react"
-import { useScrollTop } from "../../lib/hooks/use-scroll-top"
 import { CopyLinkButton } from "./copy-link-button"
 import { DetailsContent } from "./details-content"
 
@@ -12,7 +11,6 @@ export interface DetailsProps {
 
 export default function Details({ job }: DetailsProps) {
     const scrollElRef = useRef<HTMLDivElement>(null)
-    const { scrollTop } = useScrollTop(scrollElRef)
 
     const href = typeof window === "undefined" ? "" : window.location.href
