+++ src/lib/components/home/details/empty-details.tsx
