+++ src/lib/components/home/details/sections/locations.tsx
