+++ src/lib/components/home/details/sections/requirements.tsx
