+++ src/lib/components/home/details/sections/responsibilities.tsx
