+++ src/lib/components/home/mobile/details-dialog.tsx
