+++ src/lib/components/home/preview-list/preview-card/preview-card.module.scss
