+++ src/lib/components/home/top-bar/top-bar.tsx
