+++ src/lib/providers/form-provider.tsx
@@ -1,8 +1,8 @@
-import { SEARCH_FORM_DEFAULT } from "@/app/search/hooks/constants"
-import { useSearchForm } from "@/app/search/hooks/useSearchForm"
-import { SearchFormData } from "@/app/search/types"
 import { ReactNode, createContext, useContext } from "react"
 import { UseFormReturn, useForm } from "react-hook-form"
+import { SEARCH_FORM_DEFAULT } from "../components/home/search/hooks/constants"
+import { useSearchForm } from "../components/home/search/hooks/useSearchForm"
+import { SearchFormData } from "../components/home/search/types"
 
 export interface FormContextValue {
     form: UseFormReturn<SearchFormData, any, undefined>
